Move remaining code from EventSource_CoreCLR to shared partition (guarded by !CORERT)
